Initial build working on Windows.
