+++ .gitignore
@@ -8,3 +8,4 @@
 /data
 /rke2
 /build
+/rke2.exe
