+++ pkg/rke2/rke2_windows.go
@@ -1,0 +1,19 @@
+// +build windows
+
+package rke2
+
+import (
+	"github.com/urfave/cli"
+)
+
+func Server(clx *cli.Context, cfg Config) error {
+	return nil
+}
+
+func Agent(clx *cli.Context, cfg Config) error {
+	return nil
+}
+
+func EtcdSnapshot(clx *cli.Context, cfg Config) error {
+	return nil
+}
